pyproject.toml
[project]
name = "ai-chatroom-theater"
version = "0.1.0"
description = "N 個 AI 角色，一齣即興劇。坐下來看戲，或隨時插話。"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Maki Chiba", email = "maki@chiba.tw" },
]
keywords = ["ai", "chatroom", "theater", "roleplay", "multi-agent", "ollama"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment :: Simulation",
]
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.28",
    "rich>=13.0",
    "pyyaml>=6.0",
    "aioconsole>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
theater = "ai_theater.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
